Add concurrency support
1. Add Data Objects to Protobuf
2. Simplify Document Design
3. remove context
4. convert templated list to list of strings to make it compatible with protobuf
5. split list methods into declaration and definition
6. Move ID, Operation and Change to protobuf
7. Add tests for Document manipulation
8. Modify current tests to use protobuf definitions

+++ crdt_lib/include/types.h
@@ -12,7 +12,7 @@ namespace nupad {
     typedef std::string PeerId;
 
     namespace clock {
-        typedef int Tick;
+        typedef int64_t Tick;
         typedef std::unordered_map<PeerId, Tick> ClockState;
     }
 }
